Rename lamda-function-factory file
